added middleware subdirectory
